seperate the Cov2D analysis from the denoise class and merge two classes into one.
